BufferQueue: Increase max slots from 32 to 64
Increases NUM_BUFFER_SLOTS from 32 to 64 and changes the mask
returned by IGBC::getReleasedBuffers from 32 to 64 bits.

Bug: 13174352

+++ libs/gui/BufferQueue.cpp
@@ -158,7 +158,7 @@ status_t BufferQueue::consumerDisconnect() {
     return mConsumer->disconnect();
 }
 
-status_t BufferQueue::getReleasedBuffers(uint32_t* slotMask) {
+status_t BufferQueue::getReleasedBuffers(uint64_t* slotMask) {
     return mConsumer->getReleasedBuffers(slotMask);
 }
